# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "starter",
    "version": "1.0.0",
    "title": "starter",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:4000",
  "basePath": "/backend",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/electronic-components/list": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "description": "",
            "required": true
          }
        ],
        "tags": [
          "Electronic Component Api"
        ],
        "summary": "Get list components of a Project",
        "description": "",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "price": {
                  "type": "integer"
                },
                "amount": {
                  "type": "integer"
                },
                "total_cost": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "Failed",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/electronic-components/upload": {
      "post": {
        "consumes": [
          "multipart/formdata"
        ],
        "produces": [
          "multipart/formdata"
        ],
        "parameters": [
          {
            "name": "file",
            "in": "formData",
            "description": "Upload only xlsx file",
            "type": "file",
            "required": true
          },
          {
            "name": "project_id",
            "in": "query",
            "description": "Project ID to upload",
            "type": "integer",
            "required": true
          }
        ],
        "tags": [
          "Electronic Component Api"
        ],
        "summary": "Upload xlsx file to add list of electronic components",
        "description": "",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Failed",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/projects/list": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Project Api"
        ],
        "summary": "Get all Projects data",
        "description": "",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "income": {
                  "type": "integer"
                },
                "expected_start_time": {
                  "type": "string",
                  "example": "2020-01-01"
                },
                "expected_finish_time": {
                  "type": "string",
                  "example": "2020-01-01"
                },
                "real_start_time": {
                  "type": "string",
                  "example": "2020-01-01"
                },
                "real_finish_time": {
                  "type": "string",
                  "example": "2020-01-01"
                },
                "note": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string",
                  "example": "2020-01-01 12:00:00"
                },
                "total_cost": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "Failed",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/projects/find": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "type": "integer",
            "in": "query",
            "description": "Get by id (Fill in only blank)",
            "required": false
          },
          {
            "name": "title",
            "type": "string",
            "in": "query",
            "description": "Get by title (Fill in only blank)",
            "required": false
          }
        ],
        "tags": [
          "Project Api"
        ],
        "summary": "Get project data by title or id",
        "description": "",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "income": {
                  "type": "integer"
                },
                "expected_start_time": {
                  "type": "string",
                  "example": "2020-01-01"
                },
                "expected_finish_time": {
                  "type": "string",
                  "example": "2020-01-01"
                },
                "real_start_time": {
                  "type": "string",
                  "example": "2020-01-01"
                },
                "real_finish_time": {
                  "type": "string",
                  "example": "2020-01-01"
                },
                "note": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string",
                  "example": "2020-01-01 12:00:00"
                },
                "total_cost": {
                  "type": "integer"
                },
                "member": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "pay": {
                      "type": "integer"
                    }
                  }
                },
                "cost": {
                  "type": "object",
                  "properties": {
                    "server_cost": {
                      "type": "integer"
                    },
                    "incurred_cost": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/projects/create": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Project"
            }
          }
        ],
        "tags": [
          "Project Api"
        ],
        "summary": "Create new Project",
        "description": "",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "income": {
                      "type": "integer"
                    },
                    "expected_start_time": {
                      "type": "string",
                      "example": "2020-01-01"
                    },
                    "expected_finish_time": {
                      "type": "string",
                      "example": "2020-01-01"
                    },
                    "real_start_time": {
                      "type": "string",
                      "example": "2020-01-01"
                    },
                    "real_finish_time": {
                      "type": "string",
                      "example": "2020-01-01"
                    },
                    "note": {
                      "type": "string"
                    },
                    "member": {
                      "type": "object",
                      "properties": {
                        "username": {
                          "type": "string"
                        },
                        "pay": {
                          "type": "integer"
                        }
                      }
                    },
                    "cost": {
                      "type": "object",
                      "properties": {
                        "server": {
                          "type": "integer"
                        },
                        "cost_incurred": {
                          "type": "integer"
                        }
                      }
                    },
                    "id": {
                      "type": "integer"
                    },
                    "total_cost": {
                      "type": "integer"
                    },
                    "created_at": {
                      "type": "string",
                      "example": "2020-01-01 12:00:00"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/projects/edit/{id}": {
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "type": "integer",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Project"
            }
          }
        ],
        "tags": [
          "Project Api"
        ],
        "summary": "Edit existing Project",
        "description": "",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "edited project"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "income": {
                      "type": "integer"
                    },
                    "expected_start_time": {
                      "type": "string",
                      "example": "2020-01-01"
                    },
                    "expected_finish_time": {
                      "type": "string",
                      "example": "2020-01-01"
                    },
                    "real_start_time": {
                      "type": "string",
                      "example": "2020-01-01"
                    },
                    "real_finish_time": {
                      "type": "string",
                      "example": "2020-01-01"
                    },
                    "note": {
                      "type": "string"
                    },
                    "member": {
                      "type": "object",
                      "properties": {
                        "username": {
                          "type": "string"
                        },
                        "pay": {
                          "type": "integer"
                        }
                      }
                    },
                    "cost": {
                      "type": "object",
                      "properties": {
                        "server": {
                          "type": "integer"
                        },
                        "cost_incurred": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/projects/delete/{id}": {
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true
          }
        ],
        "tags": [
          "Project Api"
        ],
        "summary": "Delete existing Project",
        "description": "",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "removed project"
                }
              }
            }
          },
          "500": {
            "description": "Failed",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "post_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "create_post": "create pets in your account",
        "list_post": "list posts in your account",
        "edit_post": "modify posts in your account",
        "delete_post": "delete posts in your account"
      }
    },
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
  "definitions": {
...
    "Project": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "income": {
          "type": "integer"
        },
        "expected_start_time": {
          "type": "string",
          "example": "2020-01-01"
        },
        "expected_finish_time": {
          "type": "string",
          "example": "2020-01-01"
        },
        "real_start_time": {
          "type": "string",
          "example": "2020-01-01"
        },
        "real_finish_time": {
          "type": "string",
          "example": "2020-01-01"
        },
        "note": {
          "type": "string"
        },
        "member": {
          "type": "object",
          "properties": {
            "username": {
              "type": "string"
            },
            "pay": {
              "type": "integer"
            }
          },
          "example": [
            {
              "username": "username1",
              "pay": 0
            },
            {
              "username": "username2",
              "pay": 0
            }
          ]
        },
        "cost": {
          "type": "object",
          "properties": {
            "server": {
              "type": "integer"
            },
            "cost_incurred": {
              "type": "integer"
            }
          }
        }
      }
    },
...
  },
...
}